Added GUI and Write to Log powershell files
The GUI and Write to log files will allow you to use the main ps1 file
as a backend while keeping users from accessing it. They will only use
the GUI file to run the code. The write to log file can be used to write
to a specified log.